requirements.txt
discord.py
flask
PyNaCl
dnspython
requests
beautifulsoup4
